+++ blueprints/proxy/routes.py
@@ -13,12 +13,7 @@ import random
 from flask import jsonify, request
 
 from blueprints.proxy import proxy_bp
-from blueprints.proxy.utils import (
-    create_headers,
-    create_json_data,
-    get_copilot_token,
-    proxy_request,
-)
+from blueprints.proxy.utils import get_copilot_token, proxy_request
 from cache.cache_token import err_tokens
 from config import (
     CHAT_COMPLETION_ROUTE,
@@ -26,12 +21,10 @@ from config import (
     COMPLETION_ROUTE,
     COMPLETION_URL,
     GET_TOKEN_ROUTE,
-    GITHUB_GET_TOKEN_URL,
     GITHUB_TOKEN,
     GPT_CHAT_URL,
     GPT_KEY,
     GPT_MODEL,
-    PROXY_GPT_CHAT_COMPLETION,
     TOKEN_MAX_ERR_COUNT,
     USE_GPT_PROXY,
 )
@@ -104,28 +97,4 @@ async def proxy_copilot_chat_completion():
         res = await proxy_request(new_request, GPT_CHAT_URL, max_retry=max_retry)
     else:
         res = await proxy_request(request, CHAT_COMPLETION_URL)
-    return res
-
-
-@proxy_bp.route(f"/v1/{CHAT_COMPLETION_ROUTE}", methods=["POST"])
-@conditional_proxy_request(CHAT_COMPLETION_URL, PROXY_GPT_CHAT_COMPLETION)
-async def proxy_copilot_chat_completion_v1():
-    """
-    使用 github token 获取 copilot-chat 的提示接口
-    """
-
-    github_token = request.headers.get("Authorization", "")
-    github_token = github_token.strip("Bearer ")
-    status_code, copilot_token = await get_copilot_token(
-        github_token, GITHUB_GET_TOKEN_URL
-    )
-    if status_code != 200:
-        return jsonify(copilot_token), status_code
-    copilot_token = copilot_token.get("token")
-    headers = create_headers(copilot_token)
-    json_data, is_stream = create_json_data(request)
-    new_request = fake_request("POST", json=json_data, headers=headers)
-    res = await proxy_request(new_request, CHAT_COMPLETION_URL)
-    if is_stream:
-        res.content_type = "text/event-stream; charset=utf-8"
     return res