+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [2023-12-31]
+
+### Deleted
+
+- 删除了 `copilot to chat` 的功能，需要此功能见 [**copilot_to_chatgpt4**](https://github.com/lvguanjun/copilot_to_chatgpt4)
+
 ## [2023-12-10]
 
 ### Added